==> src/data_etl/9-poi-data-processing.ipynb <==
# POI data processing
1. Construct a postgres database called osm_de, with extra extensions installed: postgis, hstore.
2. Run the below command line.
`osm2pgsql -d osm_de -U postgres -W -H localhost -P 5433 -S D:\nine-euro-ticket-de\src\osm2pgsql-1.9.1-x64\osm2pgsql-bin\flex-config\pois.lua -O flex D:\nine-euro-ticket-de\dbs\geo\germany-latest.osm.pbf`

%load_ext autoreload
%autoreload 2
%cd D:\nine-euro-ticket-de

# Load libs
import pandas as pd
import os
os.environ['USE_PYGEOS'] = '0'
import openai
import workers
import geopandas as gpd
from tqdm.notebook import tqdm
import sqlalchemy
import time
import numpy as np

ylist = [2022, 2023]
mlist = [2, 3, 4, 5]
ty = np.random.choice(mlist, size=3)
print(ty)

# Data location
user = workers.keys_manager['database']['user']
password = workers.keys_manager['database']['password']
port = workers.keys_manager['database']['port']
db_name = workers.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

# Data location for OSM data of Sweden (April 18, 2024)
db_name_osm = workers.keys_manager['osmdb']['name']
engine_osm = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name_osm}?gssencmode=disable')

## 1. Load POI data
'historic' 'tourism' 'leisure' 'sport' 'shop' 'office' 'craft' - include all.

'amenity' - exclude some that won't have longer stay, e.g., recycle point.

'emergency'- exclude all.

# Get pois from database
gdf_pois = gpd.GeoDataFrame.from_postgis(sql="""SELECT osm_id, "class", subclass, name, geom FROM pois;""", con=engine_osm)
gdf_pois.head()

gdf_pois = gdf_pois.to_crs(4326)
gdf_pois.loc[:, 'x'] = gdf_pois.loc[:, 'geom'].x
gdf_pois.loc[:, 'y'] = gdf_pois.loc[:, 'geom'].y
gdf_pois.loc[:, 'coords'] = gdf_pois.apply(lambda row: ','.join([str(row['y']), str(row['x'])]), axis=1)

df_pois_tp = gdf_pois.groupby(['class', 'subclass']).size().to_frame(name='count').reset_index()
print(len(df_pois_tp))

print(gdf_pois['class'].unique())

df_pois_tp.sort_values(by=['count'], ascending=False).to_csv('dbs/poi/class_subclass.csv', index=False)

### 1.1 Check some classes

gdf_pois.loc[(gdf_pois['class'] == 'amenity') & (gdf_pois['subclass'] == 'proposed'), ['osm_id', 'name', 'coords']]

53.82281899915785,9.2869184

### 1.2 Mark each combination
keep = 1, include

keep = 0, exclude

## 2. Clean up class-subclass combinations

df_pois_tp = pd.read_csv('dbs/poi/class_subclass_marked.csv')
df_pois_tp = df_pois_tp.loc[df_pois_tp['keep'] == 1, :]
len(df_pois_tp), df_pois_tp['count'].sum()

### 2.1 Subclass = yes
Make the subclass same as the class.

df_pois_tp.loc[:, 'subclass'] = df_pois_tp.apply(lambda row: row['class'] \
    if row['subclass'] == 'yes' else row['subclass'], axis=1)

### 2.2 Create preliminary tags

df_pois_tp.loc[:, 'tag'] = df_pois_tp.apply(lambda row: row['class'] \
    if row['class'] in ('office', 'craft') else row['subclass'], axis=1)

df_tags = df_pois_tp.groupby(['class', 'tag'])['count'].sum().reset_index()
df_tags.sort_values(by=['tag', 'count'], ascending=False, inplace=True)

df_tags.to_csv('dbs/poi/class_tag.csv', index=False)

### 2.3 Merge, clean, and label tags
This is done manually outside this notebook.

df_tags = pd.read_csv('dbs/poi/class_tag_marked.csv')

df_pois_tp = pd.merge(df_pois_tp, df_tags[['class', 'tag', 'class_f', 'tag_f', 'remove']], on=['class', 'tag'], how='left')
df_pois_tp = df_pois_tp.loc[df_pois_tp['remove'].isna()]
df_pois_tp.drop(columns=['remove'], inplace=True)
print(len(df_pois_tp), df_pois_tp.tag_f.nunique())

# Fix two errors
df_pois_tp.loc[(df_pois_tp['class']=='leisure') &\
               (df_pois_tp['subclass']=='sports_centre'), ['class_f', 'tag_f']] = \
    ('leisure', 'sports_centre')
df_pois_tp.loc[(df_pois_tp['class']=='shop') &\
               (df_pois_tp['subclass']=='till'), ['class_f', 'tag_f']] = \
    ('shop', 'shop')

for _, g in df_pois_tp.groupby('class_f'):
    if _ == 'shop':
        print(', '.join(g['tag_f'].unique()))

df_tags_f = df_pois_tp.groupby(['class_f', 'tag_f'])['count'].sum().reset_index().sort_values(by=['class_f', 'count'], ascending=False)

### 2.4 Use GPT-4 learned categories for labeling

df_cat = pd.read_excel('dbs/poi/categories.xlsx')
df_cat.loc[:, 'label'] = df_cat.apply(lambda row: row['subcategory'] if row['type'] == 0\
    else '-'.join([row['category'], row['subcategory']]), axis=1)
df_cat.iloc[0]

# Part 1 - office, historic, craft
df_pois_tp1 = df_pois_tp.loc[df_pois_tp['class_f'].isin(['historic', 'office', 'craft']), :].copy()
df_pois_tp2 = df_pois_tp.loc[~df_pois_tp['class_f'].isin(['historic', 'office', 'craft']), :].copy()
df_pois_tp1.loc[:, 'label'] = df_pois_tp1['class_f'].apply(lambda x: x.capitalize())
df_pois_tp1.head()

# Part 2 - amenity, leisure, tourism, shop
def poi_category(x, categories_str=None):
    flag = 0
    while flag != 1:
        try:
            response = openai.ChatCompletion.create(
              model="gpt-4",
              messages=[
                {
                  "role": "system",
                  "content": f"You will be presented with points of interest tags from OpenStreetMap and your job is to provide the most suitable tag from the following list. Choose ONLY from the list of tags provided here:\n\n{categories_str}"
                },
                {
                  "role": "user",
                  "content": x
                }
              ],
              temperature=0,
              max_tokens=1024,
              top_p=1,
              frequency_penalty=0,
              presence_penalty=0
            )
            flag = 1
            cate = response.choices[0].message.content
        except:
            time.sleep(1)
    return cate

label_dict = dict()

for c in ('amenity', 'leisure', 'tourism', 'shop', 'sport'):
    if c not in label_dict:
        print(f'Working on {c}...')
        categories_str = ', '.join(df_cat.loc[df_cat.class_f.str.contains(c), 'label'].values)
        label_dict[c] = dict()
        for _, row in tqdm(df_pois_tp2.loc[df_pois_tp2['class_f']==c, :].iterrows(), 
                           desc=f'Labeling {c}'):
            if row['tag_f'] not in label_dict[c]:
                label_dict[c][row['tag_f']] = poi_category(row['tag_f'], categories_str=categories_str)
    else:
        print(f'Finished labeling {c}.')

df_pois_tp2.loc[:, 'label'] = df_pois_tp2.apply(lambda row: label_dict[row['class_f']][row['tag_f']], axis=1)
df_pois_tp2.head()

df_pois_tp_p = pd.concat([df_pois_tp1, df_pois_tp2])
df_pois_tp_p.to_csv('dbs/poi/class_f_tag_f_categories.csv', index=False)

## 3. Enrich POI data

df_pois_tp_p = pd.read_csv('dbs/poi/class_f_tag_f_categories_marked.csv')
df_pois_tp_p['label'] = df_pois_tp_p['label'].fillna(df_pois_tp_p['theme'])
df_pois_tp_p.head()

print('Before cleaning', len(gdf_pois))
gdf_pois = pd.merge(gdf_pois, 
                    df_pois_tp_p[['class', 'subclass', 'theme', 'label']], 
                    on=['class', 'subclass'],
                    how='inner')
print('After cleaning', len(gdf_pois))

gdf_pois.head()

gdf_pois[['osm_id', 'class', 'subclass', 'name', 'theme', 'label', 'geom']].\
    to_postgis('poi', engine, if_exists='replace')  # Options: 'replace', 'append', 'fail'

## 4 Check unique locations in detected stops

batch_up = 300
df_u_list = []
for batch in tqdm(range(0, batch_up), desc='Loading stops'):
    df_u_list.append(pd.read_parquet(f'dbs/stops_p/stops_p_{batch}.parquet').\
                     drop_duplicates(subset=['latitude', 'longitude'])[['latitude', 'longitude']].\
                     reset_index(drop=True))
df_u = pd.concat(df_u_list)
df_u.drop_duplicates(subset=['latitude', 'longitude'], inplace=True)
len(df_u)

df_u.loc[:, 'loc_id'] = range(0, len(df_u))
df_u.to_sql('unique_stops', engine, schema='stops', index=False, if_exists='append', method='multi', chunksize=10000)

batch = 9
df = pd.read_parquet(f'dbs/stops_p/stops_p_{batch}.parquet')
df.head()

device2group = {x:np.random.randint(1, 21) for x in list(df.device_aid.unique())}
df.loc[:, 'home2grp'] = df['device_aid'].map(device2group)
df.iloc[0]
