==> src/data_etl/13-visitation-patterns-weight.ipynb <==
# Visitation patterns
Output: `dbs/visits_day_sg/`

%load_ext autoreload
%autoreload 2
%cd D:\nine-euro-ticket-de

# Load libs
import pandas as pd
import os
os.environ['USE_PYGEOS'] = '0'
import workers
import h3
from tqdm import tqdm
import sqlalchemy
import numpy as np
from p_tqdm import p_map
import pickle

# Pyspark set up
os.environ['JAVA_HOME'] = "C:/Java/jdk-1.8"
from pyspark.sql import SparkSession
import sys
from pyspark import SparkConf
# Set up pyspark
os.environ['PYSPARK_PYTHON'] = sys.executable
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable
# Create new context
spark_conf = SparkConf().setMaster("local[18]").setAppName("MobiSeg")
spark_conf.set("spark.executor.heartbeatInterval","3600s")
spark_conf.set("spark.network.timeout","7200s")
spark_conf.set("spark.sql.files.ignoreCorruptFiles","true")
spark_conf.set("spark.driver.memory", "56g")
spark_conf.set("spark.driver.maxResultSize", "0")
spark_conf.set("spark.executor.memory","8g")
spark_conf.set("spark.memory.fraction", "0.6")
spark_conf.set("spark.sql.session.timeZone", "UTC")
spark = SparkSession.builder.config(conf=spark_conf).getOrCreate()
java_version = spark._jvm.System.getProperty("java.version")
print(f"Java version used by PySpark: {java_version}")
print('Web UI:', spark.sparkContext.uiWebUrl)

# Data location
user = workers.keys_manager['database']['user']
password = workers.keys_manager['database']['password']
port = workers.keys_manager['database']['port']
db_name = workers.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

## 1. Device filtering

df_d = pd.read_sql("""SELECT device_aid, "2019", "2022", "2023", grid_1km FROM home_g;""", con=engine)

# Share of devices that only appear in one year
for year in ('2019', '2022', '2023'):
    print(f'Share of devices with only year {year}', len(df_d.loc[df_d[year] == 100, :]) / len(df_d))

print(f'Share of devices with 2022 and 2019', len(df_d.loc[(df_d['2019'] > 0) & (df_d['2022'] > 0), :]) / len(df_d))

print(f'Share of devices with 2022 and 2023', len(df_d.loc[(df_d['2023'] > 0) & (df_d['2022'] > 0), :]) / len(df_d))

print(f'Share of devices with all years', len(df_d.loc[(df_d['2019'] > 0) & (df_d['2022'] > 0) & (df_d['2023'] > 0), :]) / len(df_d))

### 2.1 Devices sharing same home grids for three years

def year_coverage(data):
    y1, y2, y3 = data['2019'].sum(), data['2022'].sum(), data['2023'].sum()
    if (y1 > 0) & (y2 > 0) & (y3 > 0):
        return pd.Series(dict(yr_c=1))
    else:
        return pd.Series(dict(yr_c=0))

tqdm.pandas()
df_d_yc = df_d.groupby('grid_1km').progress_apply(year_coverage, include_groups=False).reset_index()

shared_grids = df_d_yc.loc[df_d_yc.yr_c == 1, 'grid_1km'].unique()
devices2keep = df_d.loc[df_d['grid_1km'].isin(shared_grids), 'device_aid'].unique()
len(devices2keep)

## 2. Find the devices having records Mar-May 2022 and 2023

data_folder = os.path.join('dbs/combined_hex2visits_day/')
paths2stops = {int(x.split('_')[-1].split('.')[0]): os.path.join(data_folder, x)\
               for x in list(os.walk(data_folder))[0][2]}
paths2stops_list = list(paths2stops.values())
print(paths2stops_list[:2])

### 2.1 Overall filtering

def devices_period_stats(file_path=None):
    import pandas as pd
    def period_stats(data):
        if len(data) > 0:
            return pd.Series(dict(no_active_days=data['date'].nunique(),
                                  no_rec=len(data),
                                  no_hex=data['h3_id'].nunique()))
        else:
            return pd.Series(dict(no_active_days=0,
                                  no_rec=0,
                                  no_hex=0))
    df_g = pd.read_parquet(file_path)
    df_g['date'] = pd.to_datetime(df_g['date'])
    # Define filtering condition (Month: 3 to 5, Year: 2022 or 2023)
    filtered_df = df_g[
        ((df_g['date'].dt.month >= 3) & (df_g['date'].dt.month <= 5)) &  # Months: March to May
        (df_g['date'].dt.year.isin([2022, 2023]))                     # Years: 2022 & 2023
    ]
    return filtered_df.groupby('device_aid').apply(period_stats).reset_index()

# Use p_map for parallel processing with progress bar
df_indi_list = p_map(devices_period_stats, paths2stops_list, num_cpus=18)
df_indi = pd.concat(df_indi_list)
print("No. of individual devices covered", len(df_indi))

df_indi.to_sql('stops_hex_indi_dt', engine, schema='data_desc', index=False, method='multi', if_exists='replace', chunksize=10000)

df_indi[['no_active_days', 'no_rec', 'no_hex']].describe()

### 2.2 Refine the coverage to make sure the devices
1) Cover 2022-03 - 2022-04 and 2022-05
2) Cover 2023-03 - 2023-04 and 2023-05

def devices_period_stats_stringent(file_path=None):
    import pandas as pd
    def period_stats(data):
        if len(data) > 0:
            return pd.Series(dict(no_active_days=data['date'].nunique(),
                                  no_rec=len(data),
                                  no_hex=data['h3_id'].nunique()))
        else:
            return pd.Series(dict(no_active_days=0,
                                  no_rec=0,
                                  no_hex=0))
    df_g = pd.read_parquet(file_path)
    df_g['date'] = pd.to_datetime(df_g['date'])
    # Define filtering condition (Month: 3 to 5, Year: 2022 or 2023)
    filtered_df = df_g[
        ((df_g['date'].dt.month >= 3) & (df_g['date'].dt.month <= 5)) &  # Months: March to May
        (df_g['date'].dt.year.isin([2022, 2023]))                     # Years: 2022 & 2023
    ].copy()
    filtered_df.loc[:, 'period'] = filtered_df['date'].dt.month.apply(lambda x: 1 if x == 5 else 0)
    filtered_df.loc[:, 'ym'] = filtered_df['date'].dt.year.astype(str) + filtered_df['period'].astype(str)
    return filtered_df.groupby(['device_aid', 'ym']).apply(period_stats, include_groups=False).reset_index()

ind_f = devices_period_stats_stringent(file_path=paths2stops_list[0])

def comp_check(data):
    if len(data) == 4:
        return pd.Series(dict(comp=1))
    return pd.Series(dict(comp=0))
ind_f = ind_f.groupby('device_aid').apply(lambda x: comp_check(x), include_groups=False).reset_index()
print(len(ind_f[ind_f['comp'] == 1]) / len(ind_f))

# Use p_map for parallel processing with progress bar
df_indi_list = p_map(devices_period_stats_stringent, paths2stops_list, num_cpus=18)
df_indi = pd.concat(df_indi_list)

df_indi = df_indi[df_indi['no_active_days'] > 0]
print("No. of individual devices covered", df_indi['device_aid'].nunique())

# Optimized function
def comp_check_fast(df):
    # Count the number of rows per group
    group_sizes = df.groupby('device_aid').size()
    # Create a Series where 1 indicates groups with 4 rows, 0 otherwise
    result = (group_sizes == 4).astype(int).rename('comp')
    return result.reset_index()

# Apply the optimized function
result = comp_check_fast(df_indi)
print(result.head())

print(len(result[result['comp'] == 1]) / len(result), len(result[result['comp'] == 1]))

devices2keep = result[result['comp'] == 1]['device_aid'].unique()

# Save the list to a file
with open('dbs/devices2keep.pkl', 'wb') as f:
    pickle.dump(devices2keep, f)

print("List saved to 'devices2keep.pkl'")

## 3. Individual weight
We only use individuals with over 15 nights at home.
### 3.1 Focus on a subset of individual devices


#devices2keep = pd.read_sql("""SELECT device_aid, no_active_days FROM data_desc.stops_hex_indi_dt;""", con=engine)
#devices2keep = devices2keep['device_aid'].to_list()

df_home = pd.read_sql("""SELECT device_aid, grid_1km, pop_1km FROM home_g;""", con=engine)
df_h = pd.read_sql("""SELECT device_aid, count FROM home WHERE count > 14;""", con=engine)
df_home = pd.merge(df_home, df_h, on='device_aid', how='left')
df_home.drop(columns=['count'], inplace=True)

# Generate weights only for individuals covered in the analysis time period
df_home = df_home.loc[df_home['device_aid'].isin(devices2keep), :]
tqdm.pandas()
df_home_s = df_home.groupby('grid_1km').progress_apply(lambda x: pd.Series(dict(count=len(x))), include_groups=False).reset_index()
df_home_s = pd.merge(df_home, df_home_s, on='grid_1km', how='left')
df_home_s.loc[:, 'wt_p'] = df_home_s.loc[:, 'pop_1km'] / df_home_s.loc[:, 'count']

w0 = ((np.std(df_home_s.loc[:, 'wt_p']) / np.mean(df_home_s.loc[:, 'wt_p'])) ** 2 + 1) ** 0.5 * 3.5 * np.median(df_home_s.loc[:, 'wt_p'])
df_home_s.loc[df_home_s['wt_p'] > w0, 'wt_p'] = w0

df_home_s[['device_aid', 'wt_p']].to_sql('weight_dt', engine, schema='public', index=False,
                                    method='multi', if_exists='replace')

df_home_s[['device_aid', 'wt_p']].to_parquet("dbs/weight_dt.parquet", index=False)

## 4. Device filtering for h3 grids

data_folder = os.path.join('dbs/combined_hex2visits_day/')
paths2stops_list = [data_folder + x for x in os.listdir(data_folder)]
paths2stops_list[:2]

paths2stops = {v.split('/')[2].split('.')[0].split('_')[1]: v for v in paths2stops_list}

df_indi = pd.read_sql("""SELECT device_aid, grdi, net_rent_100m FROM home_g;""", con=engine)
df_indi.loc[df_indi['grdi'] < 0, 'grdi'] = 0

# Only consider DT-analysis covered devices
# df_home_s = pd.read_sql("""SELECT * FROM weight_dt;""", con=engine)
df_indi = pd.merge(df_home_s[['device_aid', 'wt_p']], df_indi, on='device_aid', how='left')
grdi = df_indi['grdi'].median()
net_rent_100m = df_indi['net_rent_100m'].median()
df_indi['grdi'] = df_indi['grdi'].fillna(grdi)
df_indi['net_rent_100m'] = df_indi['net_rent_100m'].fillna(net_rent_100m)
print(f"No. of devices with grdi and net_rent_100m: {len(df_indi)}")

df_indi.dropna(inplace=True)
print(f"No. of devices with grdi and net_rent_100m: {len(df_indi)}")

h3_id_list = []
for k, v in tqdm(paths2stops.items(), desc='Adding individual group/weight to devices'):
    df = pd.read_parquet(v)
    if 'wt_p' not in df.columns:
        df = pd.merge(df, df_indi[['device_aid', 'grdi', 'net_rent_100m', 'wt_p']], on='device_aid', how='left')
        df.dropna(inplace=True)
        h3_id_list += list(df['h3_id'].unique())
        h3_id_list = list(set(h3_id_list))
        df.to_parquet(f'dbs/combined_hex2visits_day_sg_dt/stops_{k}.parquet', index=False)

df_h3 = pd.DataFrame(h3_id_list, columns=['h3_id'])
upper_reso = 3
tqdm.pandas()
df_h3.loc[:, f'h3_parent_{upper_reso}'] = df_h3['h3_id'].progress_apply(lambda x: h3.cell_to_parent(x, upper_reso))
print(df_h3[f'h3_parent_{upper_reso}'].nunique())
df_h3.to_parquet('dbs/combined_hex2visits_day_sg_dt_h3_batches.parquet', index=False)
