==> src/data_etl/15-rain-pt-access-fuel-stations.ipynb <==
# Precipitation and distance to nearest public transit station

%load_ext autoreload
%autoreload 2
%cd D:\nine-euro-ticket-de

# Load libs
import pandas as pd
import os
os.environ['USE_PYGEOS'] = '0'
import geopandas as gpd
from tqdm import tqdm
import workers
import sqlalchemy
import rasterio
import zipfile
from ftplib import FTP
import numpy as np
from sklearn.neighbors import KDTree

# Data location
user = workers.keys_manager['database']['user']
password = workers.keys_manager['database']['password']
port = workers.keys_manager['database']['port']
db_name = workers.keys_manager['database']['name']
engine = sqlalchemy.create_engine(f'postgresql://{user}:{password}@localhost:{port}/{db_name}?gssencmode=disable')

## 0. Download precipitation data

ftp = FTP('opendata.dwd.de')  # Replace with your FTP address
ftp.login()
ftp.cwd('/climate_environment/CDC/observations_germany/climate/daily/more_precip/historical')
local_directory = 'dbs/precipitation/daily'
# Function to download a file
def download_file(filename):
    local_filepath = os.path.join(local_directory, filename)
    with open(local_filepath, 'wb') as file:
        ftp.retrbinary(f"RETR {filename}", file.write)
# List files in the directory and download each file
filenames = ftp.nlst()  # List files in the directory
for filename in filenames:
    download_file(filename)

# Close the connection
ftp.quit()

## 1. Load precipitation data

file_path = 'dbs/precipitation/daily/'
station_file = 'RR_Tageswerte_Beschreibung_Stationen.txt'
paths2stations = dict()
for x in list(os.walk(file_path))[0][2]:
    if '.txt' not in x:
        paths2stations[x.split('_')[2]] = os.path.join(file_path, x)
paths2stations_list = list(paths2stations.values())

column_names = ['station_id', 'start_date', 'end_date', 
                'height', 'lng', 'lat', 'name', 'state', 'ex', 'ex2', 'ex3', 'ex4']
df_s = pd.read_csv(file_path + station_file, skiprows=2, skipinitialspace=True,
                   encoding='latin-1', header=None, delimiter=r"\s+", names=column_names)
df_s = df_s.loc[df_s.end_date > 20190501, ['station_id', 'start_date', 'end_date', 'height', 'lng', 'lat']]

df_s.head()

df_s.describe()

df_s.station_id.nunique()

### 1.1 Daily records processing

p11 = pd.to_datetime('2019050100', format='%Y%m%d%H')
p12 = pd.to_datetime('2019093000', format='%Y%m%d%H')
p21 = pd.to_datetime('2022020100', format='%Y%m%d%H')
p22 = pd.to_datetime('2022093000', format='%Y%m%d%H')
p31 = pd.to_datetime('2023020100', format='%Y%m%d%H')
p32 = pd.to_datetime('2023093000', format='%Y%m%d%H')

def rec_load(zip_file_path=None):
    # Open the zip file
    with zipfile.ZipFile(zip_file_path, 'r') as z:
        # List all files in the zip file
        file_list = z.namelist()
        # Find the file starting with 'product'
        product_file_name = next((f for f in file_list if f.startswith('produkt')), None)
        # Check if the file was found
        if product_file_name:
            # Read the file content
            with z.open(product_file_name) as f:
                # Load the content into a DataFrame
                df = pd.read_csv(f, sep=';')  # Adjust the separator as needed
                df.columns = ['station_id', 'date', 'QN_6', 'RS', 'RSF', 'SH_TAG', 'NSH_TAG', 'eor']
                return df
        else:
            return None
            
def rec_process(df=None):
    df = df.loc[df.RS != -999, ['station_id', 'date', 'RS']]
    df.loc[:, 'date_time'] = pd.to_datetime(df['date'].astype(str), format='%Y%m%d')
    # Filter the DataFrame
    mask = (
        (df['date_time'] >= p11) & (df['date_time'] <= p12) |
        (df['date_time'] >= p21) & (df['date_time'] <= p22) |
        (df['date_time'] >= p31) & (df['date_time'] <= p32)
    )
    df = df.loc[mask]
    if len(df) > 0:
        df['date_d'] = df['date_time'].dt.date
        return df
    else:
        return None

df_d_list = []
for zip_file_path in tqdm(paths2stations_list, desc='Processing daily records'):
    df = rec_load(zip_file_path=zip_file_path)
    if df is not None:
        df = rec_process(df=df)
        if df is not None:
            df_d_list.append(df)

df_day = pd.concat(df_d_list)
df_day.station_id.nunique()

### 1.2 Save data

df_day[['station_id', 'date', 'RS']].to_sql('daily', con=engine, schema='precipitation', index=False,
                                            method='multi', if_exists='replace', chunksize=5000)

df_s.to_sql('stations', con=engine, schema='precipitation', index=False,
            method='multi', if_exists='replace', chunksize=5000)

## 2. Load POI data and add elevation information

# Read the raster file
raster_path = 'dbs/geo/srtm_germany_dtm/srtm_germany_dtm.tif'
raster = rasterio.open(raster_path)

# Read the GeoDataFrame with places
gdf = gpd.GeoDataFrame.from_postgis("""SELECT osm_id, geom FROM poi;""", con=engine)
print(len(gdf))

gdf = gpd.GeoDataFrame(gdf, geometry=gdf['geom'])

coord_list = [(x, y) for x, y in zip(gdf["geometry"].x, gdf["geometry"].y)]

# Create a new column in the GeoDataFrame to store the elevation data
gdf['elevation'] = [x[0] for x in raster.sample(coord_list)]

## 3. Assign nearby weather station

df_s.head()

print('Process stations.')
gdf_s = workers.df2gdf_point(df_s, 'lat', 'lng', crs=4326, drop=False)
gdf_s = gdf_s.to_crs(25832)  # Projection in meter for Germany
gdf_s.loc[:, 'y'] = gdf_s.geometry.y
gdf_s.loc[:, 'x'] = gdf_s.geometry.x
gdf_s = gdf_s.reset_index(drop=True)
tree = KDTree(gdf_s[["y", "x"]], metric="euclidean")

print('Process POIs.')
gdf = gdf.to_crs(25832)
gdf.loc[:, 'y'] = gdf.geometry.y
gdf.loc[:, 'x'] = gdf.geometry.x
gdf.replace([np.inf, -np.inf], np.nan, inplace=True)
gdf.dropna(subset=["x", "y"], how="any", inplace=True)

print('Search for nearest POI.')
radius = 80*1000    # 80 km radius
ind, dist = tree.query_radius(gdf[["y", "x"]].to_records(index=False).tolist(),
                                   r=radius, return_distance=True, count_only=False, sort_results=True)

gdf.loc[:, 'station_num'] = [len(x) for x in ind]

stations_list = []
for x, ele in tqdm(zip(ind, gdf.elevation), 'Candidate stations'):
    sts = gdf_s.loc[x, ['station_id', 'height']]
    sts.loc[:, 'ele_diff'] = abs(sts.loc[:, 'height'] - ele)
    sta_list = [str(j) for j in sts.loc[sts.ele_diff < 150, 'station_id'].values[:15]]
    stations_list.append(','.join(sta_list))

gdf.loc[:, 'station_id'] = stations_list

gdf[['osm_id', 'elevation', 'station_num', 'station_id']].\
    to_sql('poi_station', con=engine, schema='precipitation', index=False,
           method='multi', if_exists='replace', chunksize=5000)

## 4. Load public transit data

df_stops = pd.read_csv('dbs/geo/gtfs-germany/stops.txt', delimiter=',')
df_stops = df_stops[['stop_id', 'stop_lat', 'stop_lon']].drop_duplicates(subset=['stop_lat', 'stop_lon'])

len(df_stops)

df_stops.head()

### 4.1 Find the number of public transport stations nearby
Within 800 m radius

print('Process pt stations.')
gdf_stops = workers.df2gdf_point(df_stops, 'stop_lon', 'stop_lat', crs=4326, drop=False)
gdf_stops = gdf_stops.to_crs(25832)  # Projection in meter for Germany
gdf_stops.loc[:, 'y'] = gdf_stops.geometry.y
gdf_stops.loc[:, 'x'] = gdf_stops.geometry.x
gdf_stops = gdf_stops.reset_index(drop=True)
tree = KDTree(gdf_stops[["y", "x"]], metric="euclidean")

gdf.head()

print('Search for nearest pt stations.')
radius = 800    # 800 m radius
ind, dist = tree.query_radius(gdf[["y", "x"]].to_records(index=False).tolist(),
                                   r=radius, return_distance=True, count_only=False, sort_results=True)

gdf.loc[:, 'pt_station_num'] = [len(x) for x in ind]

gdf[['osm_id', 'pt_station_num']].\
    to_sql('poi_pt_station', con=engine, schema='public_transport', index=False,
           method='multi', if_exists='replace', chunksize=5000)
df_stops.to_sql('pt_stations', con=engine, schema='public_transport', index=False,
           method='multi', if_exists='replace', chunksize=5000)

## 5. Fuel stations
Assigning fuel prices based on stations near the origin (home location).
Get five nearby stations within 5 km radius.

### 5.1 Process home locations

df_indi = pd.read_sql("""SELECT device_aid, num_unique_poi FROM home_g;""", con=engine)
df_indi = pd.merge(df_indi,
                   pd.read_sql("""SELECT device_aid, latitude, longitude FROM home;""", con=engine),
                   on='device_aid', how='left')

gdf_indi = workers.df2gdf_point(df_indi, 'latitude', 'longitude', crs=4326, drop=False)

print('Process home locations.')
gdf_indi = gdf_indi.to_crs(25832)
gdf_indi.loc[:, 'y'] = gdf_indi.geometry.y
gdf_indi.loc[:, 'x'] = gdf_indi.geometry.x
gdf_indi.replace([np.inf, -np.inf], np.nan, inplace=True)
gdf_indi.dropna(subset=["x", "y"], how="any", inplace=True)

### 5.2 Process fuel stations

df_fs = pd.read_csv('dbs/fuel_price/stations.csv')
df_fs.head()

print('Process fuel stations.')
gdf_fs = workers.df2gdf_point(df_fs[['uuid', 'latitude', 'longitude']], 'latitude', 'longitude', crs=4326, drop=False)
gdf_fs = gdf_fs.to_crs(25832)  # Projection in meter for Germany
gdf_fs.loc[:, 'y'] = gdf_fs.geometry.y
gdf_fs.loc[:, 'x'] = gdf_fs.geometry.x
gdf_fs = gdf_fs.reset_index(drop=True)

tree = KDTree(gdf_fs[["y", "x"]], metric="euclidean")

print('Search for nearest stations.')
radius = 10*1000    # 80 km radius
ind, dist = tree.query_radius(gdf_indi[["y", "x"]].to_records(index=False).tolist(),
                              r=radius, return_distance=True, count_only=False, sort_results=True)

gdf_indi.loc[:, 'station_num'] = [len(x) for x in ind]
len(gdf_indi.loc[gdf_indi['station_num'] == 0, :]) / len(gdf_indi)

stations_list = []
for x in tqdm(ind, 'Candidate stations'):
    sts = gdf_fs.loc[x, ['uuid']]
    sta_list = [str(j) for j in sts.loc[:, 'uuid'].values[:5]]
    stations_list.append(','.join(sta_list))
gdf_indi.loc[:, 'station_id'] = stations_list

gdf_indi[['device_aid', 'station_num', 'station_id']].\
    to_sql('fuel_station', con=engine, index=False, method='multi', if_exists='replace', chunksize=5000)

### 5.3 Fuel prices

file_path = 'dbs/fuel_price/'
price_records = dict()

for yr in ('2023', '2022'):
    for month in ('02', '03', '04', '05', '06', '07', '08', '09'):
        data_folder = os.path.join(file_path, yr, month)
        price_records.update({x.split('-prices')[0]: os.path.join(data_folder, x)\
                            for x in list(os.walk(data_folder))[0][2]})
yr = '2019'
for month in ('05', '06', '07', '08', '09'):
    data_folder = os.path.join(file_path, yr, month)
    price_records.update({x.split('-prices')[0]: os.path.join(data_folder, x)\
                        for x in list(os.walk(data_folder))[0][2]})
len(price_records)

def daily_station_average(data):
    data = data[['date', 'e5', 'e10']].dropna()
    return pd.Series(dict(date = data['date'].values[0].split(' ')[0],
                          e5 = data.loc[data['e5'] != 0, 'e5'].mean(),
                          e10 = data.loc[data['e10'] != 0, 'e10'].mean()))

daily_station_price = []
for k, v in tqdm(price_records.items(), desc='Averaging daily fuel price'):
    df_f = pd.read_csv(v)
    df_sp = df_f.groupby('station_uuid').apply(daily_station_average, include_groups=False).reset_index()
    df_sp.to_sql('fuel_station_price', con=engine, index=False, method='multi', if_exists='append', chunksize=5000)
